pyproject.toml
[project]
name = "dismech-warp"
version = "0.1.0"
description = "DiSMech backend for Newton"
readme = "README.md"
authors = [{ name = "ryanchaiyakul", email = "ryanchaiyakul@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "warp-lang>=1.11.1",
    "newton[sim]",
    "usd-core>=25.11",
    "socu",
]

[build-system]
requires = ["uv_build>=0.8.18,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "symengine>=0.14.1",
]

[tool.uv.sources]
newton = { path = "external/newton", editable = true }
socu = { git = "https://github.com/PREDICT-EPFL/socu.git" }

[tool.pytest.ini_options]
testpaths = ["tests"]
